test: add tests
