+++ src/index.js
@@ -3,6 +3,9 @@ import * as API from './api.js'
 import { ShardFetcher } from './shard.js'
 import * as Shard from './shard.js'
 
+/** @param {string} s */
+const isPrintableASCII = s => /^[\x20-\x7E]*$/.test(s)
+
 /**
  * Put a value (a CID) for the given key. If the key exists it's value is
  * overwritten.
@@ -16,80 +19,104 @@ import * as Shard from './shard.js'
 export const put = async (blocks, root, key, value) => {
   const shards = new ShardFetcher(blocks)
   const rshard = await shards.get(root)
+
+  if (rshard.value.keyChars !== Shard.KeyCharsASCII) {
+    throw new Error(`unsupported key character set: ${rshard.value.keyChars}`)
+  }
+  if (!isPrintableASCII(key)) {
+    throw new Error('key contains non-ASCII characters')
+  }
+  // ensure utf8 encoded key is smaller than max
+  if (new TextEncoder().encode(key).length > rshard.value.maxKeySize) {
+    throw new Error(`UTF-8 encoded key exceeds max size of ${rshard.value.maxKeySize} bytes`)
+  }
+
   const path = await traverse(shards, rshard, key)
   const target = path[path.length - 1]
-  const skey = key.slice(target.prefix.length) // key within the shard
+  const skey = key.slice(target.value.prefix.length) // key within the shard
 
   /** @type {API.ShardEntry} */
   let entry = [skey, value]
+  let targetEntries = [...target.value.entries]
 
   /** @type {API.ShardBlockView[]} */
   const additions = []
 
-  // if the key in this shard is longer than allowed, then we need to make some
-  // intermediate shards.
-  if (skey.length > target.value.maxKeyLength) {
-    const pfxskeys = Array.from(Array(Math.ceil(skey.length / target.value.maxKeyLength)), (_, i) => {
-      const start = i * target.value.maxKeyLength
-      return {
-        prefix: target.prefix + skey.slice(0, start),
-        skey: skey.slice(start, start + target.value.maxKeyLength)
+  for (const [i, e] of targetEntries.entries()) {
+    const [k, v] = e
+
+    // is this just a replace?
+    if (k === skey) break
+
+    // do we need to shard this entry?
+    const shortest = k.length < skey.length ? k : skey
+    const other = shortest === k ? skey : k
+    let common = ''
+    for (const char of shortest) {
+      const next = common + char
+      if (!other.startsWith(next)) break
+      common = next
+    }
+    if (common.length) {
+      /** @type {API.ShardEntry[]} */
+      let entries = []
+
+      // if the existing entry key or new key is equal to the common prefix,
+      // then the existing value / new value needs to persist in the parent
+      // shard. Otherwise they persist in this new shard.
+      if (common !== skey) {
+        entries = Shard.putEntry(entries, [skey.slice(common.length), value])
       }
-    })
+      if (common !== k) {
+        entries = Shard.putEntry(entries, [k.slice(common.length), v])
+      }
 
-    let child = await Shard.encodeBlock(
-      Shard.withEntries([[pfxskeys[pfxskeys.length - 1].skey, value]], target.value),
-      pfxskeys[pfxskeys.length - 1].prefix
-    )
-    additions.push(child)
-
-    for (let i = pfxskeys.length - 2; i > 0; i--) {
-      child = await Shard.encodeBlock(
-        Shard.withEntries([[pfxskeys[i].skey, [child.cid]]], target.value),
-        pfxskeys[i].prefix
+      let child = await Shard.encodeBlock(
+        Shard.withEntries(entries, { ...target.value, prefix: target.value.prefix + common })
       )
       additions.push(child)
-    }
+  
+      // need to spread as access by index does not consider utf-16 surrogates
+      const commonChars = [...common]
 
-    entry = [pfxskeys[0].skey, [child.cid]]
-  }
-
-  let shard = Shard.withEntries(Shard.putEntry(target.value.entries, entry), target.value)
-  let child = await Shard.encodeBlock(shard, target.prefix)
-
-  if (child.bytes.length > shard.maxSize) {
-    const common = Shard.findCommonPrefix(shard.entries, entry[0])
-    if (!common) throw new Error('shard limit reached')
-    const { prefix, matches } = common
-    const block = await Shard.encodeBlock(
-      Shard.withEntries(
-        matches
-          .filter(([k]) => k !== prefix)
-          .map(([k, v]) => [k.slice(prefix.length), v]),
-        shard
-      ),
-      target.prefix + prefix
-    )
-    additions.push(block)
-
-    /** @type {API.ShardEntryLinkValue | API.ShardEntryLinkAndValueValue} */
-    let value
-    const pfxmatch = matches.find(([k]) => k === prefix)
-    if (pfxmatch) {
-      if (Array.isArray(pfxmatch[1])) {
-        // should not happen! all entries with this prefix should have been
-        // placed within this shard already.
-        throw new Error(`expected "${prefix}" to be a shard value but found a shard link`)
+      // create parent shards for each character of the common prefix
+      for (let i = commonChars.length - 1; i > 0; i--) {
+        const parentConfig = { ...target.value, prefix: target.value.prefix + commonChars.slice(0, i).join('') }
+        /** @type {API.ShardEntryLinkValue | API.ShardEntryValueValue | API.ShardEntryLinkAndValueValue} */
+        let parentValue
+        // if the first iteration and the existing entry key is equal to the
+        // common prefix, then existing value needs to persist in this parent
+        if (i === commonChars.length - 1 && common === k) {
+          if (Array.isArray(v)) throw new Error('found a shard link when expecting a value')
+          parentValue = [child.cid, v]
+        } else if (i === commonChars.length - 1 && common === skey) {
+          parentValue = [child.cid, value]
+        } else {
+          parentValue = [child.cid]
+        }
+        const parent = await Shard.encodeBlock(Shard.withEntries([[commonChars[i], parentValue]], parentConfig))
+        additions.push(parent)
+        child = parent
       }
-      value = [block.cid, pfxmatch[1]]
-    } else {
-      value = [block.cid]
-    }
 
-    shard.entries = shard.entries.filter(e => matches.every(m => e[0] !== m[0]))
-    shard = Shard.withEntries(Shard.putEntry(shard.entries, [prefix, value]), shard)
-    child = await Shard.encodeBlock(shard, target.prefix)
+      // remove the sharded entry
+      targetEntries.splice(i, 1)
+
+      // create the entry that will be added to target
+      if (commonChars.length === 1 && common === k) {
+        if (Array.isArray(v)) throw new Error('found a shard link when expecting a value')
+        entry = [commonChars[0], [child.cid, v]]
+      } else if (commonChars.length === 1 && common === skey) {
+        entry = [commonChars[0], [child.cid, value]]
+      } else {
+        entry = [commonChars[0], [child.cid]]
+      }
+      break
+    }
   }
+
+  const shard = Shard.withEntries(Shard.putEntry(targetEntries, entry), target.value)
+  let child = await Shard.encodeBlock(shard)
 
   // if no change in the target then we're done
   if (child.cid.toString() === target.cid.toString()) {
@@ -98,10 +125,10 @@ export const put = async (blocks, root, key, value) => {
 
   additions.push(child)
 
-  // path is root -> shard, so work backwards, propagating the new shard CID
+  // path is root -> target, so work backwards, propagating the new shard CID
   for (let i = path.length - 2; i >= 0; i--) {
     const parent = path[i]
-    const key = child.prefix.slice(parent.prefix.length)
+    const key = child.value.prefix.slice(parent.value.prefix.length)
     const value = Shard.withEntries(
       parent.value.entries.map((entry) => {
         const [k, v] = entry
@@ -112,7 +139,7 @@ export const put = async (blocks, root, key, value) => {
       parent.value
     )
 
-    child = await Shard.encodeBlock(value, parent.prefix)
+    child = await Shard.encodeBlock(value)
     additions.push(child)
   }
 
@@ -133,7 +160,7 @@ export const get = async (blocks, root, key) => {
   const rshard = await shards.get(root)
   const path = await traverse(shards, rshard, key)
   const target = path[path.length - 1]
-  const skey = key.slice(target.prefix.length) // key within the shard
+  const skey = key.slice(target.value.prefix.length) // key within the shard
   const entry = target.value.entries.find(([k]) => k === skey)
   if (!entry) return
   return Array.isArray(entry[1]) ? entry[1][1] : entry[1]
@@ -153,7 +180,7 @@ export const del = async (blocks, root, key) => {
   const rshard = await shards.get(root)
   const path = await traverse(shards, rshard, key)
   const target = path[path.length - 1]
-  const skey = key.slice(target.prefix.length) // key within the shard
+  const skey = key.slice(target.value.prefix.length) // key within the shard
 
   const entryidx = target.value.entries.findIndex(([k]) => k === skey)
   if (entryidx === -1) return { root, additions: [], removals: [] }
@@ -192,13 +219,13 @@ export const del = async (blocks, root, key) => {
     }
   }
 
-  let child = await Shard.encodeBlock(shard, path[path.length - 1].prefix)
+  let child = await Shard.encodeBlock(shard)
   additions.push(child)
 
   // path is root -> shard, so work backwards, propagating the new shard CID
   for (let i = path.length - 2; i >= 0; i--) {
     const parent = path[i]
-    const key = child.prefix.slice(parent.prefix.length)
+    const key = child.value.prefix.slice(parent.value.prefix.length)
     const value = Shard.withEntries(
       parent.value.entries.map((entry) => {
         const [k, v] = entry
@@ -209,7 +236,7 @@ export const del = async (blocks, root, key) => {
       parent.value
     )
 
-    child = await Shard.encodeBlock(value, parent.prefix)
+    child = await Shard.encodeBlock(value)
     additions.push(child)
   }
 
@@ -236,7 +263,7 @@ export const entries = async function * (blocks, root, options = {}) {
     /** @returns {AsyncIterableIterator<API.ShardValueEntry>} */
     async function * ents (shard) {
       for (const entry of shard.value.entries) {
-        const key = shard.prefix + entry[0]
+        const key = shard.value.prefix + entry[0]
 
         if (Array.isArray(entry[1])) {
           if (entry[1][1]) {
@@ -272,7 +299,7 @@ export const entries = async function * (blocks, root, options = {}) {
               continue
             }
           }
-          yield * ents(await shards.get(entry[1][0], key))
+          yield * ents(await shards.get(entry[1][0]))
         } else {
           if (
             (prefix && key.startsWith(prefix)) ||
@@ -302,7 +329,7 @@ const traverse = async (shards, shard, key) => {
   for (const [k, v] of shard.value.entries) {
     if (key === k) return [shard]
     if (key.startsWith(k) && Array.isArray(v)) {
-      const path = await traverse(shards, await shards.get(v[0], shard.prefix + k), key.slice(k.length))
+      const path = await traverse(shards, await shards.get(v[0]), key.slice(k.length))
       return [shard, ...path]
     }
   }
