--- src/v1/api.js
@@ -1,1 +1,0 @@
-export {}
